remove base from naming when directly inherit from base. Required fix for cases where constraint has .cpp

+++ wolf_scripts/generic_func/functions.sh
@@ -131,42 +131,62 @@ createHCPPFromTemplates()
   mv "${TEMPLATES_PATH}"/tmp2.cpp "$NAME_CPP_PATH"
 }
 
+addAutodiffSpecifics()
+{
+echo "TODO"
+    # Count number of parameters
+ 
+  #const T* const _p1
+  
+    # Add () operator
+    #sed -i "/\[base class inherited methods\]/a\ \n\        \/\*\* \brief : compute the residual from the state blocks being iterated by the solver.\n \        \*\*\/\n\        template<typename T>\n\        bool operator ()(aaaa) const;\n" "$NAME_H_PATH"
+    
+#            template<typename T>
+#        bool operator ()(const T* const _p1, const T* const _o1, const T* const _p2, const T* const _o2,
+#   
+}
+
 fillWithBaseVirtualMethods()
 {
-  # Get base class	
-  H_TXT=$(cat $BASE_H_PATH)
-  H_TXT="${H_TXT##*class $BASECLASSNAME :}"
-  BASECLASS_TXT="${H_TXT%%\n\};*}"
+  if [[ $BASECLASSNAME =~ .*ConstraintAutodiff*. ]] ;
+  then
+    addAutodiffSpecifics
+  else
+    # Get base class	
+    H_TXT=$(cat $BASE_H_PATH)
+    H_TXT="${H_TXT##*class $BASECLASSNAME :}"
+    BASECLASS_TXT="${H_TXT%%\n\};*}"
 
-  echo "class $BASECLASSNAME :$BASECLASS_TXT };" > "${WOLF_SCRIPTS_PATH}"/class.h
+    echo "class $BASECLASSNAME :$BASECLASS_TXT };" > "${WOLF_SCRIPTS_PATH}"/class.h
 
-  # H file Get Virtual function declarations with help
-  sed -e '/./{H;$!d;}' -e 'x;/ = 0;/!d;' "${WOLF_SCRIPTS_PATH}"/class.h > "${WOLF_SCRIPTS_PATH}"/tmp.h
-  sed -r 's/'" = 0"'//g' "${WOLF_SCRIPTS_PATH}"/tmp.h > "${WOLF_SCRIPTS_PATH}"/tmp2.h
-  sed -i -e "/\[base class inherited methods\]/r ${WOLF_SCRIPTS_PATH}/tmp2.h" "$NAME_H_PATH" 
-  rm "${WOLF_SCRIPTS_PATH}"/tmp.h	
-  rm ${WOLF_SCRIPTS_PATH}/tmp2.h
+    # H file Get Virtual function declarations with help
+    sed -e '/./{H;$!d;}' -e 'x;/ = 0;/!d;' "${WOLF_SCRIPTS_PATH}"/class.h > "${WOLF_SCRIPTS_PATH}"/tmp.h
+    sed -r 's/'" = 0"'//g' "${WOLF_SCRIPTS_PATH}"/tmp.h > "${WOLF_SCRIPTS_PATH}"/tmp2.h
+    sed -i -e "/\[base class inherited methods\]/r ${WOLF_SCRIPTS_PATH}/tmp2.h" "$NAME_H_PATH" 
+    rm "${WOLF_SCRIPTS_PATH}"/tmp.h	
+    rm ${WOLF_SCRIPTS_PATH}/tmp2.h
     
-  # CPP file  
-  FuncInBase=$(grep -e " = 0;" -e "=0;" "${WOLF_SCRIPTS_PATH}/class.h")
-  D=";"   #Multi Character Delimiter
-  FuncList=($(echo $FuncInBase | sed -e 's/'"$D"'/\n/g' | while read line; do echo $line | sed 's/[\t ]/'"$D"'/g'; done))
+    # CPP file  
+    FuncInBase=$(grep -e " = 0;" -e "=0;" "${WOLF_SCRIPTS_PATH}/class.h")
+    D=";"   #Multi Character Delimiter
+    FuncList=($(echo $FuncInBase | sed -e 's/'"$D"'/\n/g' | while read line; do echo $line | sed 's/[\t ]/'"$D"'/g'; done))
 
-  for (( idx = $((${#FuncList[@]}-1)); idx > -1; idx-- )); do
-  	TMP=$(echo ${FuncList[idx]} | sed -r 's/'"$D"'/ /g')
-    TMP=$(echo $TMP | sed -r 's/'"virtual"'/ /g')
-    TMP=$(echo $TMP | sed -r 's/'"=0"'/ /g')
-    TMP=$(echo $TMP | sed -r 's/'" = 0"'/ /g')
-    TXTH=${TMP%$TMP##*[![:space:]]}
-    TXTCPP_3=$(echo $TXTH | sed 's/.*(//g')
-    TXTCPP_2=$(echo $TXTH | sed 's/(.*//g' | sed 's/.* //g')
-    TXTCPP_1=$(echo $TXTH | sed 's/'"$TXTCPP_2"'.*//g')
+    for (( idx = $((${#FuncList[@]}-1)); idx > -1; idx-- )); do
+  	  TMP=$(echo ${FuncList[idx]} | sed -r 's/'"$D"'/ /g')
+      TMP=$(echo $TMP | sed -r 's/'"virtual"'/ /g')
+      TMP=$(echo $TMP | sed -r 's/'"=0"'/ /g')
+      TMP=$(echo $TMP | sed -r 's/'" = 0"'/ /g')
+      TXTH=${TMP%$TMP##*[![:space:]]}
+      TXTCPP_3=$(echo $TXTH | sed 's/.*(//g')
+      TXTCPP_2=$(echo $TXTH | sed 's/(.*//g' | sed 's/.* //g')
+      TXTCPP_1=$(echo $TXTH | sed 's/'"$TXTCPP_2"'.*//g')
  
-    # CPP file
-    sed -i "/\[base class inherited methods\]/a ${TXTCPP_1}${CLASSNAME}::${TXTCPP_2}(${TXTCPP_3}\n\{\n\}\n" "$NAME_CPP_PATH"  
-  done    
+      # CPP file
+      sed -i "/\[base class inherited methods\]/a ${TXTCPP_1}${CLASSNAME}::${TXTCPP_2}(${TXTCPP_3}\n\{\n\}\n" "$NAME_CPP_PATH"  
+    done    
     
-  rm "${WOLF_SCRIPTS_PATH}"/class.h
+    rm "${WOLF_SCRIPTS_PATH}"/class.h
+  fi
 }
 
 updateCMakeLists()
